The ding sound fixed :'))
