+++ PharmacyQueue/Pharmacy/Form1.resx
